% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agro_fishery_in_out}
\alias{agro_fishery_in_out}
\alias{우리나라}
\alias{농수산물의}
\alias{수입과}
\alias{수출}
\alias{현황}
\title{agro_fishery_in_out : 초등사회 6-2}
\format{
변수 3개와 5행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{년도}}{character 2005년부터 2009년까지, 단위: 년}
  \item{\code{수입액}}{double 농수산물을 수입한 금액}
  \item{\code{수출액}}{double 농수산물을 수출한 금액}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
agro_fishery_in_out
}
\description{
초등사회 6-2 교과서에 수록된 우리나라 농수산물의 수입과 수출 현황 데이터
}
\details{
초등사회 6-2 교과서에 수록된 2005년부터 2009년까지 농수산물을 수입 및 수출한 금액 데이터
}
\keyword{datasets}
