% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_monthly_precipitation}
\alias{area_monthly_precipitation}
\alias{지역의}
\alias{월별}
\alias{강수량}
\title{area_monthly_precipitation : 초등수학 4-2}
\format{
변수 2개와 4행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{월}}{double 10월부터 1월까지 매달, 단위: 월}
  \item{\code{강수량}}{double 월 평균 강수량}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
area_monthly_precipitation
}
\description{
초등수학 4-2 교과서에 수록된 지역의 월별 강수량 데이터
}
\details{
초등수학 4-2 교과서에 수록된 10월부터 1월까지 월 평균 강수량 데이터
}
\keyword{datasets}
