% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseball_spectators_num}
\alias{baseball_spectators_num}
\alias{프로야구}
\alias{관중}
\alias{수}
\title{baseball_spectators_num : 초등수학 4-2}
\format{
변수 2개와 8행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{년}}{double 2001년부터 2008년까지 매년, 단위: 년}
  \item{\code{관중수}}{double 한 해동안 프로야구를 보러온 총 관중 수, 단위: 만명}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
baseball_spectators_num
}
\description{
초등수학 4-2 교과서에 수록된 프로야구 관중 수 데이터
}
\details{
초등수학 4-2 교과서에 수록된 2001년부터 2008년까지 매년 기록된 한 해동안 프로야구를 보러온 총 관중 수 데이터
}
\keyword{datasets}
