% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CO2_emissions}
\alias{CO2_emissions}
\alias{국가별}
\alias{이산화탄소}
\alias{배출량}
\title{CO2_emissions : 초등과학 6-2}
\format{
변수 4개와 10행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{국가}}{character 대한민국, 중국, 미국 등 10개 국가의 이름}
  \item{\code{이산화탄소배출량}}{double 국가별 이산화탄소 배출량}
  \item{\code{순위}}{double 국가별 순위}
  \item{\code{1인당배출량}}{double 국가별 1인당 이산화탄소 배출량}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
CO2_emissions
}
\description{
2009개정 초등과학 6-2 교과서에 수록된 국가별 이산화탄소 배출량 데이터
}
\details{
초등과학 6-2 교과서에 수록된 국가별 이산화탄소 배출량, 순위, 1인당 이산화탄소 배출량 데이터
}
\keyword{datasets}
