% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosmos_height_change}
\alias{cosmos_height_change}
\alias{코스모스}
\alias{키의}
\alias{변화}
\title{cosmos_height_change : 초등수학 4-2}
\format{

}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
cosmos_height_change
}
\description{
초등수학 4-2 교과서에 수록된 코스모스 키의 변화 데이터
}
\details{
초등수학 4-2 교과서에 수록된 월요일부터 일요일까지 매일 잰 코스모스의 키 데이터
}
\keyword{datasets}
