% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counseling_user_num}
\alias{counseling_user_num}
\alias{월별}
\alias{상담실}
\alias{이용자수}
\title{counseling_user_num : 초등수학 5-2}
\format{
변수 2개와 5행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{월}}{double 6월부터 10월까지 매월, 단위: 월}
  \item{\code{이용자}}{double 상담실을 이용한 사람의 수, 단위: 명}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
counseling_user_num
}
\description{
초등수학 5-2 교과서에 수록된 월별 상담실 이용자수 데이터
}
\details{
초등수학 5-2 교과서에 수록된 6월부터 10월까지 매월 상담실을 이용한 사람의 수 데이터
}
\keyword{datasets}
