% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_urbanization_rate}
\alias{country_urbanization_rate}
\alias{국가별}
\alias{도시화율}
\alias{변화}
\title{country_urbanization_rate : 초등수학 5-2}
\format{

}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
country_urbanization_rate
}
\description{
초등수학 5-2 교과서에 수록된 국가별 도시화율 변화 데이터
}
\details{
초등수학 5-2 교과서에 수록된 1985년부터 2015년까지 5년 주기로 측정한 한국, 일본, 중국, 미국의 도시화율 데이터
}
\keyword{datasets}
