% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{day_lowest_temp}
\alias{day_lowest_temp}
\alias{하루}
\alias{중}
\alias{최저기온}
\title{day_lowest_temp : 초등수학 4-2}
\format{
변수 2개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{날짜}}{double 1일부터 6일까지 매일, 단위: 일}
  \item{\code{최고기온}}{double 하루 중 가장 낮았던 기온, 단위: 섭씨}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
day_lowest_temp
}
\description{
초등수학 4-2 교과서에 수록된 하루 중 최저기온 데이터
}
\details{
초등수학 4-2 교과서에 수록된 1일부터 6일까지 매일 측정된 하루 중 가장 낮았던 기온 데이터
}
\keyword{datasets}
