% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df10.R
\name{df10}
\alias{df10}
\title{Resumo de Dados}
\format{
Um \code{data.frame} com xx observacoes e 2 variaveis, em que

\describe{

\item{\code{Nome da Variavel 1}}{Descricao da variavel}

\item{\code{Nome da Variavel 1}}{Descricao da variavel}

}
}
\source{
Referencia Bibliografica
}
\description{
Descrever a atividade.
}
\examples{
#Fazer a leitura do Banco de Dados
data(df10)
#Para visualizar a Estrutura e os tipos de variaveis do Banco de dados faca:
str(df10)
#Para visualizar os nomes das colunas do banco de dados faca:
names(df10)
#'#Caso queira ver um resumo do conjunto de dados:
summary(df10)

}
\keyword{Digite}
\keyword{chave}
\keyword{palavras}
\keyword{tres}
