% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distance_time_graph}
\alias{distance_time_graph}
\alias{거리-시간}
\alias{그래프}
\title{distance_time_graph : 초등과학 5-2}
\format{
변수 2개와 7행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{거리}}{double 거리, 단위: cm}
  \item{\code{시간}}{double 거리에 따른 시간, 단위: 초}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
distance_time_graph
}
\description{
초등과학 5-2 교과서에 수록된 거리-시간 그래프 데이터
}
\details{
초등과학 5-2 교과서에 수록된 거리에 따른 시간을 나타낸 그래프 데이터
}
\keyword{datasets}
