% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{export_amount}
\alias{export_amount}
\alias{대한민국}
\alias{수출액}
\title{export_amount : 초등사회 5-2}
\format{
변수 2개와 11행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{연도}}{double 1960년부터 2010년까지 5년 주기, 단위: 년}
  \item{\code{수출액}}{double 대한민국이 해외에 수출한 금액, 단위: 억 달러}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
export_amount
}
\description{
초등사회 5-2 교과서에 수록된 대한민국 수출액 데이터
}
\details{
초등사회 5-2 교과서에 수록된 1960년부터 2010년까지 대한민국의 해외 수출액 데이터
}
\keyword{datasets}
