% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expt_data_salt_hour}
\alias{expt_data_salt_hour}
\alias{시간별}
\alias{소금양에}
\alias{따른}
\alias{실험자료}
\title{expt_data_salt_hour : 초등수학 5-2}
\format{
변수 6개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{구분}}{character 경과된 날짜, 단위: 일}
  \item{\code{소금0g}}{double 소금이 0g일 때 경과된 날짜에 따른 데이터}
  \item{\code{소금10g}}{double 소금이 10g일 때 경과된 날짜에 따른 데이터}
  \item{\code{소금20g}}{double 소금이 20g일 때 경과된 날짜에 따른 데이터}
  \item{\code{소금30g}}{double 소금이 30g일 때 경과된 날짜에 따른 데이터}
  \item{\code{소금40g}}{double 소금이 40g일 때 경과된 날짜에 따른 데이터}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
expt_data_salt_hour
}
\description{
초등수학 5-2 교과서에 수록된 시간별 소금양에 따른 실험자료 데이터
}
\details{
초등수학 5-2 교과서에 수록된 경과된 날짜와 소금의 양에 따른 데이터
}
\keyword{datasets}
