% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farming_num}
\alias{farming_num}
\alias{농사짓는}
\alias{가구수}
\title{farming_num : 초등수학 5-2}
\format{
변수 2개와 4행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{연도}}{double 1967년부터 1997년까지 10년 주기, 단위: 년}
  \item{\code{가구수}}{double 농사짓는 가구의 수, 단위: 가구}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
farming_num
}
\description{
초등수학 5-2 교과서에 수록된 농사짓는 가구수 데이터
}
\details{
초등수학 5-2 교과서에 수록된 1967년부터 1997년까지 10년 주기로 측정한 농사짓는 가구의 수 데이터
}
\keyword{datasets}
