% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{favorite_folk_game}
\alias{favorite_folk_game}
\alias{좋아하는}
\alias{민속놀이}
\title{favorite_folk_game : 초등수학 4-2}
\format{
변수 2개와 5행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{민속놀이종목}}{character 연날리기, 윷놀이 등 민속놀이 종목의 이름}
  \item{\code{학생수}}{double 각 민속놀이를 좋아하는 학생의 수, 단위: 명}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
favorite_folk_game
}
\description{
초등수학 4-2 교과서에 수록된 좋아하는 민속놀이 데이터
}
\details{
초등수학 4-2 교과서에 수록된 연날리기, 윷놀이, 제기차기, 팽이치기, 비사치기를 좋아하는 학생의 수 데이터
}
\keyword{datasets}
