% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flower_tree_height}
\alias{flower_tree_height}
\alias{꽃나무의}
\alias{키}
\title{flower_tree_height : 초등수학 4-2}
\format{
변수 2개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{월}}{double 3월부터 8월까지, 단위: 월}
  \item{\code{꽃나무키}}{double 지면으로부터 잰 꽃나무의 키, 단위: cm}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
flower_tree_height
}
\description{
초등수학 4-2 교과서에 수록된 꽃나무의 키 데이터
}
\details{
초등수학 4-2 교과서에 수록된 3월부터 8월까지 매월 잰 꽃나무의 키 데이터
}
\keyword{datasets}
