% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gas_in_air}
\alias{gas_in_air}
\alias{공기}
\alias{중에}
\alias{포함된}
\alias{여러가지}
\alias{기체}
\title{gas_in_air : 초등과학 6-2}
\format{

}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
gas_in_air
}
\description{
초등과학 6-2 교과서에 수록된 공기 중에 포함된 여러가지 기체 데이터
}
\details{
초등과학 6-2 교과서에 수록된 공기 중에 포함된 기체의 함유량 데이터
}
\keyword{datasets}
