% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rawdata}
\alias{get_rawdata}
\alias{데이터파일가져오기}
\title{소스 데이터 파일 가져오기}
\usage{
get_rawdata(dest_dir = ".")
}
\arguments{
\item{dest_dir}{character. 파일을 가져올 디렉토리 경로.
기본값은 "."으로 현재 디렉토리}
}
\description{
초중등 통계 데이터셋을 만든 데이터 파일을 가져온다.
메뉴에서 해당 파일 이름을 선택하면, 지정한 경로에 파일이 복사된다.
}
\examples{
\dontrun{
# 현재 작업 경로에 데이터 파일을 복사한다.
get_rawdata()

# c:/temp에 선택한 데이터 파일을 복사한다.
get_rawdata("c:/temp")
}

}
