% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{global_warming_gas}
\alias{global_warming_gas}
\alias{지구온난화의}
\alias{원인}
\alias{기체}
\title{global_warming_gas : 초등과학 6-2}
\format{

}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
global_warming_gas
}
\description{
초등과학 6-2 교과서에 수록된 지구온난화의 원인 기체 데이터
}
\details{
초등과학 6-2 교과서에 수록된 지구온난화의 원인 기체의 영향 정도 데이터
}
\keyword{datasets}
