% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ground_water_temp}
\alias{ground_water_temp}
\alias{지면과}
\alias{수면의}
\alias{온도}
\alias{차이}
\title{ground_water_temp : 초등과학 6-2}
\format{
변수 2개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{시각}}{double 08시부터 다음날 07시까지 매시 정각}
  \item{\code{지면온도}}{double 매시 정각 지면의 온도, 단위: 섭씨}
  \item{\code{수면온도}}{double 매시 정각 수면의 온도, 단위: 섭씨}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
ground_water_temp
}
\description{
초등과학 6-2 교과서에 수록된 지면과 수면의 온도 차이 데이터
}
\details{
초등과학 6-2 교과서에 수록된 하루동안 매시 정각에 측정된 지면과 수면의 온도 데이터
}
\keyword{datasets}
