% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{high_speed_internet}
\alias{high_speed_internet}
\alias{초고속}
\alias{인터넷}
\alias{가입자수}
\title{high_speed_internet : 초등사회 5-2}
\format{
변수 2개와 10행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{연도}}{character 1998년부터 2009년까지, 단위: 년}
  \item{\code{가입자수}}{double 초고속 인터넷에 가입한 사람의 수, 단위: 명}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
high_speed_internet
}
\description{
초등사회 5-2 교과서에 수록된 초고속 인터넷 가입자수 데이터
}
\details{
초등사회 5-2 교과서에 수록된 1998년부터 2009년까지 초고속 인터넷에 가입한 사람의 수 데이터
}
\keyword{datasets}
