% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hobby_participating_num}
\alias{hobby_participating_num}
\alias{취미}
\alias{종목별}
\alias{참가}
\alias{학생}
\alias{수}
\title{hobby_participating_num : 초등수학 4-2}
\format{
변수 2개와 5행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{취미종목}}{character 합창, 연극 등 취미 종목의 이름}
  \item{\code{학생수}}{double 각 취미 종목에 참가한 학생의 수, 단위: 명}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
hobby_participating_num
}
\description{
초등수학 4-2 교과서에 수록된 취미 종목별 참가 학생 수 데이터
}
\details{
초등수학 4-2 교과서에 수록된 합창, 연극, 무용, 기악연주, 장기자랑 취미 종목에 참가한 학생의 수 데이터
}
\keyword{datasets}
