% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ice_melt_color}
\alias{ice_melt_color}
\alias{색소의}
\alias{색깔에}
\alias{따라}
\alias{얼음이}
\alias{녹은양}
\title{ice_melt_color : 초등과학 6-2}
\format{
변수 2개와 8행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{얼음색깔}}{character 얼음의 색깔}
  \item{\code{녹은양}}{double 일정 시간 후 얼음이 녹은양}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
ice_melt_color
}
\description{
초등과학 6-2 교과서에 수록된 색소의 색깔에 따라 얼음이 녹은양 데이터
}
\details{
초등과학 6-2 교과서에 수록된 얼음에 포함된 색소의 색깔에 따른 얼음의 녹은양 데이터
}
\keyword{datasets}
