% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jinwook_unit_score}
\alias{jinwook_unit_score}
\alias{진욱이의}
\alias{단원평가}
\alias{점수}
\title{jinwook_unit_score : 초등수학 5-2}
\format{
변수 2개와 5행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{단원}}{character 1단원부터 5단원까지}
  \item{\code{점수}}{double 진욱이의 단원평가 점수, 단위: 점}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
jinwook_unit_score
}
\description{
초등수학 5-2 교과서에 수록된 진욱이의 단원평가 점수 데이터
}
\details{
초등수학 5-2 교과서에 수록된 1단원부터 5단원까지 각 단원에 대한 진욱이의 단원평가 점수 데이터
}
\keyword{datasets}
