% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{junseong_typing_practice}
\alias{junseong_typing_practice}
\alias{준성이의}
\alias{타자연습}
\title{junseong_typing_practice : 초등수학 4-2}
\format{
변수 2개와 5행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{주}}{double 1주차부터 5주차까지 매주, 단위: 주차}
  \item{\code{타수}}{double 준성이의 최고 타수, 단위: 타}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
junseong_typing_practice
}
\description{
초등수학 4-2 교과서에 수록된 준성이의 타자연습 데이터
}
\details{
초등수학 4-2 교과서에 수록된 1주차부터 5주차까지 매주 기록한 준성이의 최고 타수 데이터
}
\keyword{datasets}
