% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kicks_num}
\alias{kicks_num}
\alias{제기차기}
\alias{횟수}
\title{kicks_num : 초등수학 5-2}
\format{
변수 1개와 30행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{횟수}}{double 제기차기 횟수}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
kicks_num
}
\description{
초등수학 5-2 교과서에 수록된 제기차기 횟수 데이터
}
\details{
초등수학 5-2 교과서에 수록된 제기차기 횟수 데이터
}
\keyword{datasets}
