% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{livingroom_temps}
\alias{livingroom_temps}
\alias{거실의}
\alias{온도}
\title{livingroom_temps : 초등수학 4-2}
\format{
변수 2개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{시각}}{double 4시부터 9시까지 매시 정각, 단위: 시}
  \item{\code{온도}}{double 매시 정각에 측정한 거실의 온도, 단위: 섭씨}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
livingroom_temps
}
\description{
초등수학 4-2 교과서에 수록된 거실의 온도 데이터
}
\details{
초등수학 4-2 교과서에 수록된 4시부터 9시까지 매시 정각에 측정한 거실의 온도 데이터
}
\keyword{datasets}
