% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minsu_height}
\alias{minsu_height}
\alias{민수의}
\alias{몸무게}
\title{minsu_height : 초등수학 4-2}
\format{
변수 2개와 4행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{학년}}{character 1학년부터 4학년까지}
  \item{\code{몸무게}}{double 민수의 몸무게, 단위: kg}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
minsu_height
}
\description{
초등수학 4-2 교과서에 수록된 민수의 몸무게 데이터
}
\details{
초등수학 4-2 교과서에 수록된 1학년부터 4학년까지 매학년 측정한 민수의 몸무게 데이터
}
\keyword{datasets}
