% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newborn_num}
\alias{newborn_num}
\alias{신생아수}
\title{newborn_num : 초등수학 5-2}
\format{
변수 2개와 5행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{연도}}{double 2000년부터 2008년까지 2년 주기, 단위: 년}
  \item{\code{신생아수}}{double 한 해 동안 태어난 신생아의 수, 단위: 명}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
newborn_num
}
\description{
초등수학 5-2 교과서에 수록된 신생아수 데이터
}
\details{
초등수학 5-2 교과서에 수록된 2000년부터 2008년까지 2년 주기로 측정된 한 해 동안 태어난 신생아의 수 데이터
}
\keyword{datasets}
