% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{openstat_tbl}
\alias{openstat_tbl}
\alias{데이터셋}
\title{openstat_tbl : 초중등 통계 데이터셋}
\format{
변수 2개와 50행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{name}}{character 파일명}
  \item{\code{content}}{tibble 티블 데이터프레임}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
openstat_tbl
}
\description{
초중등 교과서에 언급된 데이터셋 모음.
}
\details{
초중등 교과서에 언급된 데이터셋 모음
}
\keyword{datasets}
