% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planet_distance}
\alias{planet_distance}
\alias{행성의}
\alias{상대적}
\alias{거리}
\title{planet_distance : 초등과학 5-2}
\format{
변수 2개와 8행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{명칭}}{character 태양계 존재하는 행성 이름}
  \item{\code{거리}}{double 지구를 1로 뒀을 때 행성까지의 거리}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
planet_distance
}
\description{
초등과학 5-2 교과서에 수록된 태양계에서 행성까지의 상대적 거리 데이터
}
\details{
초등과학 5-2 교과서에 수록된 지구를 1로 뒀을 때 태양으로부터 떨어진 행성의 거리 데이터
}
\keyword{datasets}
