% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{planet_radius}
\alias{planet_radius}
\alias{행성반지름}
\title{planet_radius : 초등과학 5-2}
\format{
변수 2개와 9행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{행성이름}}{character 태양계 존재하는 행성 이름}
  \item{\code{반지름}}{double 지구를 1로 뒀을 때 행성의 반지름}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
planet_radius
}
\description{
초등과학 5-2 교과서에 수록된 지구를 1로 뒀을 때 행성의 반지름 데이터
}
\details{
초등과학 5-2 교과서에 수록된 행성의 크기에 때한 실증 데이터
}
\keyword{datasets}
