% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_age}
\alias{population_age}
\alias{연령별}
\alias{인구}
\alias{구성비}
\title{population_age : 초등사회 4-2}
\format{
변수 4개와 9행을 갖는 티블 데이터프레임:
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
population_age
}
\description{
초등사회 4-2 교과서에 수록된 연령별 인구 구성비 변화 데이터
}
\details{
초등사회 4-2 교과서에 수록된 1970년부터 2050년까지 세 그룹으로 나눈 연령별 인구 구성비 데이터
}
\keyword{datasets}
