% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{privacy_infringement}
\alias{privacy_infringement}
\alias{개인정보침해}
\alias{신고건수}
\title{privacy_infringement : 초등사회 6-2}
\format{
변수 2개와 5행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{년도}}{character 2004년부터 2008년까지, 단위: 년}
  \item{\code{건수}}{double 개인정보침해 신고가 접수된 건수, 단위: 건}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
privacy_infringement
}
\description{
초등사회 6-2 교과서에 수록된 개인정보침해 신고건수 데이터
}
\details{
초등사회 6-2 교과서에 수록된 2004년부터 2008년까지 개인정보침해 신고가 접수된 건수 데이터
}
\keyword{datasets}
