% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rice_consumption_person}
\alias{rice_consumption_person}
\alias{일제시대}
\alias{1인당}
\alias{쌀}
\alias{소비량}
\title{rice_consumption_person : 초등사회 5-2}
\format{
변수 2개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{연도}}{character 1920년부터 1930년까지 2년 주기}
  \item{\code{1인당소비량}}{double 1인당 소비한 쌀의 양}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
rice_consumption_person
}
\description{
초등사회 5-2 교과서에 수록된 일제시대 1인당 쌀 소비량 데이터
}
\details{
초등사회 5-2 교과서에 수록된 일제시대(1920~1930)에 1인당 소비한 쌀의 양 데이터
}
\keyword{datasets}
