% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rice_consumption_world}
\alias{rice_consumption_world}
\alias{쌀}
\alias{소비량}
\alias{국제}
\alias{비교}
\title{rice_consumption_world : 초등사회 6-2}
\format{
변수 3개와 3행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{국가}}{character 국가의 이름: 대한민국, 일본, 타이완}
  \item{\code{1999}}{double 1999년의 쌀 소비량}
  \item{\code{2008}}{double 2008년의 쌀 소비량}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
rice_consumption_world
}
\description{
초등사회 6-2 교과서에 수록된 쌀 소비량 국제 비교 데이터
}
\details{
초등사회 6-2 교과서에 수록된 1999년과 2008년 대한민국, 일본, 타이완의 쌀 소비량 데이터
}
\keyword{datasets}
