% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rice_produced_taken}
\alias{rice_produced_taken}
\alias{일제시대}
\alias{쌀}
\alias{생산량과}
\alias{수탈량}
\title{rice_produced_taken : 초등사회 5-2}
\format{
변수 3개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{연도}}{character 1920년부터 1930년까지 2년 주기}
  \item{\code{쌀생산량}}{double 생산한 쌀의 양}
  \item{\code{쌀수탈량}}{double 일제가 수탈한 쌀의 양}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
rice_produced_taken
}
\description{
초등사회 5-2 교과서에 수록된 일제시대 쌀 생산량과 수탈량 자료 데이터
}
\details{
초등사회 5-2 교과서에 수록된 일제시대(1920~1930)에 생산한 쌀의 양과 수탈당한 쌀의 양 데이터
}
\keyword{datasets}
