% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{running_record}
\alias{running_record}
\alias{달리기}
\alias{기록}
\title{running_record : 초등과학 5-2}
\format{
변수 2개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{레인}}{double 1부터 6까지 레인의 번호}
  \item{\code{걸린시간}}{double 달리기에 걸린 시간, 단위: 초}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
running_record
}
\description{
초등과학 5-2 교과서에 수록된 달리기 기록 데이터
}
\details{
초등과학 5-2 교과서에 수록된 레인별 달리기에 걸린시간 데이터
}
\keyword{datasets}
