% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sea_pollution}
\alias{sea_pollution}
\alias{해역오염실태자료}
\title{sea_pollution : 초등과학 3-2}
\format{
변수 3개와 16행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{연도}}{double 1993년부터 2008년까지 해역 오염실태 조사 연도}
  \item{\code{건수}}{double 해역 오염 건수}
  \item{\code{유출량}}{double 해역 오염 유출량, 단위: kl}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
sea_pollution
}
\description{
초등과학 3-2 교과서에 수록된 우리나라 해역 오염실태자료 데이터다.
}
\details{
초등과학 3-2 교과서에 수록된 1993년부터 2008년까지 우리나라 해역 오염실태자료 데이터다.
}
\keyword{datasets}
