% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seoul_daegu_temp}
\alias{seoul_daegu_temp}
\alias{서울과}
\alias{대구의}
\alias{기온변화}
\title{seoul_daegu_temp : 초등과학 6-2}
\format{
변수 3개와 24행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{시각}}{character 01시부터 24시까지 매시 정각}
  \item{\code{서울}}{double 매시 정각 서울의 기온, 단위: 섭씨}
  \item{\code{대구}}{double 매시 정각 대구의 기온, 단위: 섭씨}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
seoul_daegu_temp
}
\description{
초등과학 6-2 교과서에 수록된 서울과 대구의 기온변화 데이터
}
\details{
초등과학 6-2 교과서에 수록된 하루동안 매시 정각에 측정된 서울과 대구의 기온 데이터
}
\keyword{datasets}
