% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situp_record}
\alias{situp_record}
\alias{윗몸일으키기}
\alias{기록}
\title{situp_record : 초등수학 4-2}
\format{
변수 2개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{요일}}{character 월요일부터 토요일까지 매일, 단위: 요일}
  \item{\code{횟수}}{double 윗몸일으키기를 실시한 횟수, 단위: 회}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
situp_record
}
\description{
초등수학 4-2 교과서에 수록된 윗몸일으키기 기록 데이터
}
\details{
초등수학 4-2 교과서에 수록된 월요일부터 토요일까지 매일 윗몸일으키기를 실시한 횟수 데이터
}
\keyword{datasets}
