% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{statdata}
\alias{statdata}
\alias{datadata-package}
\title{초중등 통계 데이터셋 모음}
\description{
`statdata` 팩키지는 초중등 통계, 수학, 공학 교과서에 수록된
다양한 데이터를 포함하고 있다.
`statdata` 팩키지에 수록된 데이터를 활용하여 통계 리터러시
함양에 필요한 단순작업을 크게 줄여주고 다양한 분석기법 공유에
기여할 것으로 기대된다.
}
