% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students_height}
\alias{students_height}
\alias{학생들}
\alias{키}
\title{students_height : 초등수학 4-2}
\format{
변수 2개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{이름}}{character 학생들의 이름}
  \item{\code{키}}{double 학생들의 키, 단위: cm}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
students_height
}
\description{
초등수학 4-2 교과서에 수록된 학생들 키 데이터
}
\details{
초등수학 4-2 교과서에 수록된 6명의 학생들의 키 데이터
}
\keyword{datasets}
