% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students_math_score}
\alias{students_math_score}
\alias{학생들의}
\alias{수학}
\alias{점수}
\title{students_math_score : 초등수학 5-2}
\format{
변수 1개와 40행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{점수}}{double 학생들의 수학 점수}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
students_math_score
}
\description{
초등수학 5-2 교과서에 수록된 학생들의 수학 점수 데이터
}
\details{
초등수학 5-2 교과서에 수록된 학생들의 수학 점수 데이터
}
\keyword{datasets}
