% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students_weight}
\alias{students_weight}
\alias{학생들}
\alias{몸무게}
\title{students_weight : 초등수학 4-2}
\format{
변수 2개와 8행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{이름}}{character 학생들의 이름}
  \item{\code{몸무게}}{double 학생들의 몸무게, 단위: kg}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
students_weight
}
\description{
초등수학 4-2 교과서에 수록된 학생들 몸무게 데이터
}
\details{
초등수학 4-2 교과서에 수록된 8명의 학생들의 몸무게 데이터
}
\keyword{datasets}
