% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transportation}
\alias{transportation}
\alias{교통수단}
\title{transportation : 초등과학 5-2}
\format{
변수 2개와 4행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{교통수단}}{character 자전거, 자동차, 배, 기차 중 하나의 교통수단}
  \item{\code{이동거리}}{double 교통수단에 따른 이동거리, 단위: km}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
transportation
}
\description{
초등과학 5-2 교과서에 수록된 교통수단 데이터
}
\details{
초등과학 5-2 교과서에 수록된 교통수단별 이동거리 데이터
}
\keyword{datasets}
