% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urbanization}
\alias{urbanization}
\alias{국가별}
\alias{도시화율}
\title{urbanization : 초등사회 5-2}
\format{

}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
urbanization
}
\description{
초등사회 5-2 교과서에 수록된 국가별 도시화율 변화 데이터
}
\details{
초등사회 5-2 교과서에 수록된 1985년부터 2015년까지 한국, 일본, 중국, 미국의 도시화율 데이터
}
\keyword{datasets}
