% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_height}
\alias{water_height}
\alias{수면의}
\alias{높이}
\title{water_height : 초등수학 4-2}
\format{
변수 2개와 6행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{날짜}}{double 1일부터 6일까지 매일, 단위: 일}
  \item{\code{수면의높이}}{double 매일 측정한 수면의 높이}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
water_height
}
\description{
초등수학 4-2 교과서에 수록된 수면의 높이 데이터
}
\details{
초등수학 4-2 교과서에 수록된 1일부터 6일까지 매일 측정한 수면의 높이 데이터
}
\keyword{datasets}
