% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_temp_boric_acid}
\alias{water_temp_boric_acid}
\alias{물의}
\alias{온도와}
\alias{붕산}
\alias{녹는양}
\title{water_temp_boric_acid : 초등과학 5-2}
\format{
변수 2개와 4행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{물의온도}}{double 물의 온도, 단위: 섭씨}
  \item{\code{붕산녹는양}}{double 물의 온도에 따른 붕산이 녹는양, 단위: g}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
water_temp_boric_acid
}
\description{
초등과학 5-2 교과서에 수록된 물의 온도와 붕산 녹는양 데이터
}
\details{
초등과학 5-2 교과서에 수록된 물의 온도에 따른 붕산 녹는양 데이터
}
\keyword{datasets}
