% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_temp_height}
\alias{water_temp_height}
\alias{물의}
\alias{온도와}
\alias{물기둥}
\alias{높이}
\title{water_temp_height : 초등과학 5-2}
\format{
변수 4개와 4행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{물의온도}}{double 물의 온도, 단위: 섭씨}
  \item{\code{물기둥높이1회}}{double 첫 번째로 측정한 물의 온도에 따른 물기둥 높이}
  \item{\code{물기둥높이2회}}{double 두 번째로 측정한 물의 온도에 따른 물기둥 높이}
  \item{\code{물기둥높이3회}}{double 세 번째로 측정한 물의 온도에 따른 물기둥 높이}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
water_temp_height
}
\description{
초등과학 5-2 교과서에 수록된 물의 온도와 물기둥 높이 데이터
}
\details{
초등과학 5-2 교과서에 수록된 3회 측정된 물의 온도에 따른 물기둥 높이 데이터
}
\keyword{datasets}
