% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_temp_liquid_height}
\alias{water_temp_liquid_height}
\alias{물의}
\alias{온도와}
\alias{액체기둥}
\alias{높이}
\title{water_temp_liquid_height : 초등과학 5-2}
\format{
변수 4개와 4행을 갖는 티블 데이터프레임:
\describe{
  \item{\code{물온도}}{double 물의 온도, 단위: 섭씨}
  \item{\code{식초기둥높이}}{double 물의 온도에 따른 식초기둥 높이}
  \item{\code{우유기둥높이}}{double 물의 온도에 따른 우유기둥 높이}
  \item{\code{유리세정제기둥높이}}{double 물의 온도에 따른 유리세정제기둥 높이}
}
}
\source{
\url{http://estat.me/estat/eStat/}
}
\usage{
water_temp_liquid_height
}
\description{
초등과학 5-2 교과서에 수록된 물의 온도와 액체기둥 높이 데이터
}
\details{
초등과학 5-2 교과서에 수록된 물의 온도에 따른 식초기둥, 우유기둥, 유리세정제기둥 높이 데이터
}
\keyword{datasets}
